+++ backend/src/test/java/moaon/backend/article/dto/CreatedAtCursorTest.java
@@ -3,6 +3,7 @@ package moaon.backend.article.dto;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.time.LocalDateTime;
+import moaon.backend.global.cursor.CreatedAtCursor;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
