+++ backend/src/test/java/moaon/backend/article/dto/CursorParserTest.java
@@ -6,6 +6,10 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 
 import java.time.LocalDateTime;
 import moaon.backend.article.domain.ArticleSortBy;
+import moaon.backend.global.cursor.ClickCursor;
+import moaon.backend.global.cursor.CreatedAtCursor;
+import moaon.backend.global.cursor.Cursor;
+import moaon.backend.global.cursor.CursorParser;
 import moaon.backend.global.exception.custom.CustomException;
 import moaon.backend.global.exception.custom.ErrorCode;
 import org.junit.jupiter.api.DisplayName;
