+++ backend/src/test/java/moaon/backend/article/repository/CustomizedArticleRepositoryImplTest.java
@@ -8,13 +8,13 @@ import moaon.backend.article.domain.Article;
 import moaon.backend.article.domain.ArticleCategory;
 import moaon.backend.article.domain.ArticleSortBy;
 import moaon.backend.article.dto.ArticleQueryCondition;
-import moaon.backend.article.dto.ClickCursor;
-import moaon.backend.article.dto.CreatedAtCursor;
 import moaon.backend.fixture.ArticleFixtureBuilder;
 import moaon.backend.fixture.ArticleQueryConditionBuilder;
 import moaon.backend.fixture.Fixture;
 import moaon.backend.fixture.RepositoryHelper;
 import moaon.backend.global.config.QueryDslConfig;
+import moaon.backend.global.cursor.ClickCursor;
+import moaon.backend.global.cursor.CreatedAtCursor;
 import moaon.backend.techStack.domain.TechStack;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
