+++ backend/src/test/java/moaon/backend/fixture/ArticleQueryConditionBuilder.java
@@ -3,7 +3,7 @@ package moaon.backend.fixture;
 import java.util.List;
 import moaon.backend.article.domain.ArticleSortBy;
 import moaon.backend.article.dto.ArticleQueryCondition;
-import moaon.backend.article.dto.Cursor;
+import moaon.backend.global.cursor.Cursor;
 
 public class ArticleQueryConditionBuilder {
